requirements.txt
os
Pillow
